requirements.txt
atomicwrites==1.2.1
attrs==18.2.0
cachetools==3.0.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
Flask==1.0.2
google-api-python-client==1.7.7
google-auth==1.6.2
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.2.0
httplib2==0.12.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
more-itertools==5.0.0
oauthlib==3.0.0
pluggy==0.8.1
py==1.7.0
pyasn1==0.4.5
pyasn1-modules==0.2.3
pytest==4.1.1
python-dateutil==2.7.5
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
rsa==4.0
six==1.12.0
uritemplate==3.0.0
urllib3==1.24.1
Werkzeug==0.14.1
